+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 ## [Unreleased]
+### Changed
+- Directly parse file when adding it to the agenda files. Therefore the events are directly displayed.
 
 ## [1.0.2] - 2026-04-27
 ### Changed